+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.13.0] - 2025-02-26
 
 ### Added
+- Add posibility to set start to yesterday. [\#1047](https://github.com/destination-earth-digital-twins/Deode-Workflow/pull/1047#pullrequestreview-2677820249) (@kastelecn)
 - Print deode version and config file used for each task [\#1011](https://github.com/destination-earth-digital-twins/Deode-Workflow/pull/1011) (@uandrae)
 - Configure dependabot for frequent dependency updates. [\#989](https://github.com/destination-earth-digital-twins/Deode-Workflow/pull/989) (@observingClouds)